Rename mpl as matplotlib
