+++ histogrammar/plot/matplotlib.py
